API server with test solved and dockerfile
