+++ .github/workflows/deploy.yaml
@@ -1,26 +1,26 @@
-name: Run LINTERS when pushed
-
-on:
-  push:
-    branches: ["test"]
-
-jobs:
-  push_and_check_linters:
-    name: Linters Check
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Set up Go
-        uses: actions/setup-go@v5
-        with:
-          go-version: '1.24'
-
-      - name: Run golangci-lint
-        uses: golangci/golangci-lint-action@v7
-        with:
-          version: v2.1.6
-
-
-
+#name: Run LINTERS when pushed
+#
+#on:
+#  push:
+#    branches: ["test"]
+#
+#jobs:
+#  push_and_check_linters:
+#    name: Linters Check
+#    runs-on: ubuntu-latest
+#    steps:
+#      - name: Checkout code
+#        uses: actions/checkout@v4
+#
+#      - name: Set up Go
+#        uses: actions/setup-go@v5
+#        with:
+#          go-version: '1.24'
+#
+#      - name: Run golangci-lint
+#        uses: golangci/golangci-lint-action@v7
+#        with:
+#          version: v2.1.6
+#
+#
+#
