+++ c.out
@@ -1,0 +1,40 @@
+mode: set
+API_Basics/main.go:12.59,16.16 3 1
+API_Basics/main.go:16.16,22.3 4 1
+API_Basics/main.go:40.25,43.20 2 0
+API_Basics/main.go:43.20,46.3 2 0
+API_Basics/main.go:49.74,53.16 3 1
+API_Basics/main.go:53.16,59.3 4 0
+API_Basics/main.go:61.2,66.16 3 1
+API_Basics/main.go:66.16,72.3 4 1
+API_Basics/main.go:74.2,78.49 3 1
+API_Basics/main.go:81.67,85.16 3 1
+API_Basics/main.go:85.16,91.3 4 1
+API_Basics/main.go:93.2,93.32 1 1
+API_Basics/main.go:93.32,94.24 1 1
+API_Basics/main.go:94.24,95.12 1 1
+API_Basics/main.go:98.3,103.17 4 1
+API_Basics/main.go:103.17,108.4 3 1
+API_Basics/main.go:110.3,110.9 1 1
+API_Basics/main.go:113.2,113.36 1 1
+API_Basics/main.go:116.68,119.32 2 1
+API_Basics/main.go:119.32,121.17 2 1
+API_Basics/main.go:121.17,126.4 3 0
+API_Basics/main.go:128.3,130.17 2 1
+API_Basics/main.go:130.17,135.4 3 1
+API_Basics/main.go:139.72,143.16 3 1
+API_Basics/main.go:143.16,149.3 4 1
+API_Basics/main.go:151.2,151.32 1 1
+API_Basics/main.go:151.32,152.24 1 1
+API_Basics/main.go:152.24,153.12 1 1
+API_Basics/main.go:156.3,160.9 3 1
+API_Basics/main.go:163.2,163.36 1 1
+API_Basics/main.go:166.70,170.16 3 1
+API_Basics/main.go:170.16,176.3 4 1
+API_Basics/main.go:178.2,178.32 1 1
+API_Basics/main.go:178.32,179.24 1 1
+API_Basics/main.go:179.24,185.4 3 1
+API_Basics/main.go:188.2,188.36 1 1
+API_Basics/main.go:191.13,197.77 4 0
+API_Basics/main.go:197.77,200.3 2 0
+API_Basics/main.go:201.2,214.33 6 0
